Docs: norm
1.함수 길이를 제외한 norm양식

resolved #164
